no-caps/requirements.txt
fastapi
uvicorn
boto3
celery
whisper
pyannote.audio
python-dotenv
redis